refactor: reorg datamod dir
